updates to back end files, added model for standards
