--- controllers/api/adminRoutes.js
@@ -1,1 +1,0 @@
-// bonus: admin routes
