+++ controllers/api/standardRoutes.js
@@ -1,85 +1,42 @@
-// note this is for standard routes, may need changes
+// for bonus: + admin routes 
+// standard routes to get, post, delete (if applicable)
 
 const router = require('express').Router();
-const { User , Standard } = require('../../models');
+const { Standard } = require('../../models');
 const withAuth = require('../../utils/auth');
 
-router.get('/', async (req, res) => {
+router.post('/', withAuth, async (req, res) => {
   try {
-    // Get all projects and JOIN with user data
-    const standardData = await Standard.findAll({
-      include: [
-        {
-          model: User,
-          attributes: ['name'],
-        },
-      ],
+    //   if we chose to let users create standards
+    const newStandard = await Standard.create({
+      ...req.body,
+      user_id: req.session.user_id,
     });
 
-    // Serialize data so the template can read it
-    const standard = standardData.map((standard) => standard.get({ plain: true }));
-
-    // Pass serialized data and session flag into template
-    res.render('homepage', { 
-      standard, 
-      logged_in: req.session.logged_in 
-    });
+    res.status(200).json(newProject);
   } catch (err) {
-    res.status(500).json(err);
+    res.status(400).json(err);
   }
 });
 
-
-router.get('/standard/:id', async (req, res) => {
+router.delete('/:id', withAuth, async (req, res) => {
   try {
-    const standardData = await Standard.findByPk(req.params.id, {
-      include: [
-        {
-          model: User,
-          attributes: ['name'],
-        },
-      ],
+    const standardData = await Standard.destroy({
+      where: {
+        id: req.params.id,
+        user_id: req.session.user_id,
+      },
     });
 
-    const standard = standardData.get({ plain: true });
+    if (!projectData) {
+      res.status(404).json({ message: 'No standard found with this id!' });
+      return;
+    }
 
-    res.render('standard', {
-      ...standard,
-      logged_in: req.session.logged_in
-    });
+    res.status(200).json(projectData);
   } catch (err) {
     res.status(500).json(err);
   }
-});
-
-// Use withAuth middleware to prevent access to route
-router.get('/profile', withAuth, async (req, res) => {
-  try {
-    // Find the logged in user based on the session ID
-    const userData = await User.findByPk(req.session.user_id, {
-      attributes: { exclude: ['password'] },
-      include: [{ model: Project }],
-    });
-
-    const user = userData.get({ plain: true });
-
-    res.render('profile', {
-      ...user,
-      logged_in: true
-    });
-  } catch (err) {
-    res.status(500).json(err);
-  }
-});
-
-router.get('/login', (req, res) => {
-  // If the user is already logged in, redirect the request to another route
-  if (req.session.logged_in) {
-    res.redirect('/profile');
-    return;
-  }
-
-  res.render('login');
 });
 
 module.exports = router;
